requirements.txt
fastapi
uvicorn
groq
redis
pydantic
sse-starlette
python-dotenv
tenacity